Possible Modal Approach
I can't get my head around how the modal will change depending on what button is pressed. The idea is that the create button (which will be put at the top of the screen) would open a modal with the create component filling it inside and the update button would also be able to open a modal and be populated with the update component. Then the modal component can be put inside the read component to make it easier to access the function of getData or to access the state of the read component so that it can then update.